[TNT-217] feat: 다양한 수정

+++ src/main/java/com/tnt/infrastructure/mysql/repository/trainee/DietRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.tnt.domain.trainee.Diet;
 
-public interface DietRepository extends JpaRepository<Diet, Integer> {
+public interface DietRepository extends JpaRepository<Diet, Long> {
 
 }
